requirements.txt
anthropic>=0.15.0
pydantic>=2.0.0
pydantic-ai>=0.6.0
langgraph>=0.0.19
logfire>=0.15.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
httpx>=0.25.0
starlette>=0.27.0
jinja2>=3.1.2
